Move stories to stories directory
